Re-export p4rs crate, and use specific dep.
Downstream consumers (read: propolis) of softnpu utilize the p4rs types
in their interactions with this crate.  To ensure that the types match
up, p4rs is re-exported from this library, and the dependency locked to
a git revision. (As opposed to relying on Cargo.lock, which is useless
when built as a dependency.)